requirements.txt
fastapi
torch
torchvision
pillow
uvicorn
gdown
python-multipart

